Update to gwdetchar-whistles
I updated gwdetchar-whistles to output an html page containing all of the plots and bins that are necessary.  In order to do so, I made some edits to the whistles.py file to make it easier to reference the plots that it created.  I renamed the whistle.py file and kept the original one.